+++ activerecord/lib/active_record/fixtures.rb
@@ -395,6 +395,22 @@ module ActiveRecord
   #     <<: *DEFAULTS
   #
   # Any fixture labeled "DEFAULTS" is safely ignored.
+  #
+  # == Support to custom model class
+  #
+  # You can also set the model class in your fixtures YAML file.
+  # This is helpful when fixtures are being loaded outside tests and
+  # you cannot use +set_fixture_class+, e.g., when running
+  # <tt>rake db:fixtures:load</tt>.
+  #
+  # To load the fixtures file `accounts.yml` as the `User` model, use:
+  #
+  #   _fixture:
+  #     model_class: User
+  #   david:
+  #     name: David
+  #
+  # Any fixture labeled "_fixture" is safely ignored.
   class FixtureSet
     #--
     # An instance of FixtureSet is normally stored in a single YAML file and
@@ -578,12 +594,16 @@ module ActiveRecord
       @name     = name
       @path     = path
       @config   = config
-      @model_class = nil
 
-      if class_name.is_a?(Class) # TODO: Should be an AR::Base type class, or any?
-        @model_class = class_name
-      else
-        @model_class = class_name.safe_constantize if class_name
+      self.model_class = class_name
+
+      @fixtures = read_fixture_files(path)
+
+      # Sets the custom model class to all loaded fixtures.
+      fixture_settings = @fixtures.delete('_fixture')
+      if fixture_settings && (custom_model_class = fixture_settings['model_class'])
+        self.model_class = custom_model_class
+        @fixtures.values.each { |fixture| fixture.model_class = self.model_class }
       end
 
       @connection  = connection
@@ -591,8 +611,6 @@ module ActiveRecord
       @table_name = ( model_class.respond_to?(:table_name) ?
                       model_class.table_name :
                       self.class.default_fixture_table_name(name, config) )
-
-      @fixtures = read_fixture_files path, @model_class
     end
 
     def [](x)
@@ -761,7 +779,15 @@ module ActiveRecord
         @column_names ||= @connection.columns(@table_name).collect(&:name)
       end
 
-      def read_fixture_files(path, model_class)
+      def model_class=(class_name)
+        if class_name.is_a?(Class) # TODO: Should be an AR::Base type class, or any?
+          @model_class = class_name
+        else
+          @model_class = class_name.safe_constantize if class_name
+        end
+      end
+
+      def read_fixture_files(path)
         yaml_files = Dir["#{path}/{**,*}/*.yml"].select { |f|
           ::File.file?(f)
         } + [yaml_file_path(path)]
@@ -790,7 +816,8 @@ module ActiveRecord
     class FormatError < FixtureError #:nodoc:
     end
 
-    attr_reader :model_class, :fixture
+    attr_reader :fixture
+    attr_accessor :model_class
 
     def initialize(fixture, model_class)
       @fixture     = fixture